Added custom validator for Nicaraguan Phone Number format and isolated Phone Number test
