+++ test/LoanManager.Application.Test/Customer/Commands/Update/CommandValidatorTest.cs
@@ -44,64 +44,6 @@ public class CommandValidatorTest : BaseHandler
             .WithErrorMessage(RequiredFirstNameMessage)
             .Only();
     }
-    
-    [Fact]
-    public async Task Should_Fail_With_Empty_Phone()
-    {
-        // arrange
-        var command = CommandFaker.UpdateCommand() with
-        {
-            Phone = string.Empty
-        };
-
-        // act
-        var sut = await _validator.TestValidateAsync(command);
-
-        // assert
-        sut.ShouldHaveValidationErrorFor(x => x.Phone)
-            .WithErrorMessage(RequiredPhoneNumberMessage);
-        sut.ShouldHaveValidationErrorFor(x => x.Phone)
-            .WithErrorMessage(InvalidPhoneNumberFormatMessage);
-        sut.ShouldHaveValidationErrorFor(x => x.Phone)
-            .WithErrorMessage(InvalidPhoneNumberLengthMessage);
-    }
-    
-    [Fact]
-    public async Task Should_Fail_With_Invalid_Format_PhoneNumber()
-    {
-        // arrange
-        var command = CommandFaker.UpdateCommand() with
-        {
-            Phone = _faker.Phone.PhoneNumber("#######Y")
-        };
-        
-        // act
-        var sut = await _validator.TestValidateAsync(command);
-
-        // assert
-        sut.ShouldHaveValidationErrorFor(x => x.Phone)
-            .WithErrorMessage(InvalidPhoneNumberFormatMessage)
-            .Only();
-    }
-    
-    [Fact]
-    public async Task Should_Fail_With_Invalid_Character_Length_PhoneNumber()
-    {
-        // arrange
-        var command = CommandFaker.UpdateCommand() with
-        {
-            Phone = _faker.Phone.PhoneNumber("####")
-        };
-        
-        // act
-        var sut = await _validator.TestValidateAsync(command);
-
-        // assert
-        sut.ShouldHaveValidationErrorFor(x => x.Phone)
-            .WithErrorMessage(InvalidPhoneNumberFormatMessage);
-        sut.ShouldHaveValidationErrorFor(x => x.Phone)
-            .WithErrorMessage(InvalidPhoneNumberLengthMessage);
-    }
 
     [Fact]
     public async Task Should_Succeed_With_FirstName_LastName_PhoneNumber_CustomerId()
